chore: add user to /models
