+++ packages/admin/src/models/User.ts
@@ -1,0 +1,7 @@
+export type User = {
+    firstName: string;
+    lastName: string;
+    email: string;
+    id: string;
+    token: string;
+};
